Link Kobim's fork of sqs-insight in README.md
I was very confused at how this repository was using a "DynamicEndpoints" key in the configuration when the linked project doesn't have that option. Then I took a look at the `Dockerfile`. 

There's more stuff in the README.md that should be updated since it doesn't seem to be aware of the whole dynamic endpoints thing which is nice!